Few follow-ups on comments
